moved fixture down
